backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
pymysql==1.1.1
mysqlclient>=2.2.1
cryptography==41.0.7
python-decouple==3.8
django-filter==23.5
